Removed redundant call to doubleValue().
git-svn-id: http://chianti.ucsd.edu/svn@26217 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ core3/core-task-impl/trunk/src/main/java/org/cytoscape/task/internal/zoom/ZoomEdit.java
@@ -22,8 +22,7 @@ final class ZoomEdit extends CyAbstractEdit {
 	public void redo() {
 		super.redo();
 
-		final double currentScaleFactor =
-			view.getVisualProperty(NETWORK_SCALE_FACTOR).doubleValue();
+		final double currentScaleFactor = view.getVisualProperty(NETWORK_SCALE_FACTOR);
 		view.setVisualProperty(NETWORK_SCALE_FACTOR, currentScaleFactor * factor);
 		view.updateView();
 	}
@@ -31,8 +30,7 @@ final class ZoomEdit extends CyAbstractEdit {
 	public void undo() {
 		super.undo();
 
-		final double currentScaleFactor =
-			view.getVisualProperty(NETWORK_SCALE_FACTOR).doubleValue();
+		final double currentScaleFactor = view.getVisualProperty(NETWORK_SCALE_FACTOR);
 		view.setVisualProperty(NETWORK_SCALE_FACTOR, currentScaleFactor / factor);
 		view.updateView();
 	}
